@@ -1,37 +1,41 @@
 {
-  "name": "tatarozhdez",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "node-sass": "^4.14.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "proxy": "http://localhost:3000"
+    "name": "tatarozhdez",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "animated-scroll-to": "2.3.0",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "react-particles": "2.4.0",
+        "react-router-dom": "6.4.2",
+        "react-scripts": "5.0.1",
+        "sass": "1.55.0",
+        "tsparticles": "2.4.0"
+    },
+    "scripts": {
+        "start": "react-scripts start",
+        "build": "react-scripts build"
+    },
+    "eslintConfig": {
+        "extends": "react-app"
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    },
+    "proxy": "http://localhost:3000",
+    "devDependencies": {
+        "eslint": "8.26.0",
+        "eslint-config-prettier": "8.5.0",
+        "eslint-plugin-react": "7.31.10",
+        "prettier": "2.7.1"
+    }
 }
